requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-dynamic-db-router==0.3.0
djangorestframework==3.9.1
mysqlclient==1.4.1
numpy==1.18.1
pandas==1.0.1
pip==19.0.3
PyMySQL==0.9.3
setuptools==40.8.0
sqlparse==0.3.0
urllib3==1.24.2